Rename chrono.Chrono to global.chrono
