+++ src/main/java/org/bytedeco/javacpp/global/chrono.java
@@ -1,7 +1,7 @@
-package org.bytedeco.javacpp.chrono;
+package org.bytedeco.javacpp.global;
 
 import org.bytedeco.javacpp.annotation.Properties;
 
 @Properties(inherit = org.bytedeco.javacpp.presets.chrono.class)
-public class Chrono {
+public class chrono {
 }
