+++ src/main/java/org/bytedeco/javacpp/presets/chrono.java
@@ -1,5 +1,6 @@
 package org.bytedeco.javacpp.presets;
 
+import org.bytedeco.javacpp.annotation.Platform;
 import org.bytedeco.javacpp.annotation.Properties;
 import org.bytedeco.javacpp.tools.Info;
 import org.bytedeco.javacpp.tools.InfoMap;
@@ -8,7 +9,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
 @Properties(
     inherit = javacpp.class,
     target = "org.bytedeco.javacpp.chrono",
-    global = "org.bytedeco.javacpp.chrono.Chrono"
+    global = "org.bytedeco.javacpp.global.chrono"
 )
 public class chrono implements InfoMapper {
     @Override public void map(InfoMap infoMap) {
