+++ src/main/java9/module-info.java
@@ -3,6 +3,7 @@ module org.bytedeco.javacpp {
     exports org.bytedeco.javacpp;
     exports org.bytedeco.javacpp.annotation;
     exports org.bytedeco.javacpp.chrono;
+    exports org.bytedeco.javacpp.global;
     exports org.bytedeco.javacpp.indexer;
     exports org.bytedeco.javacpp.tools;
     exports org.bytedeco.javacpp.presets;
